backend/requirements.txt
fastapi
uvicorn
supabase
requests
python-dotenv
pydantic
pydantic-settings
